Add files via upload
The student Record System is based on the concept of managing student records. There’s a login system available for this system, the user can freely use its feature. This mini project contains limited features, but the essential one. This project is helpful for managing student information by adding, updating, removing, viewing, and searching for details.